refactor: findPath의 예외를 ControllerAdvice에서 처리
